+++ src/View/PlaceholderSupport.java
@@ -1,4 +1,4 @@
-package guis;
+package View;
 
 import javax.swing.*;
 import java.awt.*;
